feat: refactor [params] to [entityId] (#424)
